chore: renamed api test config folder
