+++ src/api/ecommerce-cdc/steps/ecommerce_cdc_steps.py
@@ -12,7 +12,7 @@ from pathlib import Path
 
 from behave import given, then, when
 
-HELPERS_DIR = Path(__file__).resolve().parents[3] / "utility" / "api_test" / "checkout"
+HELPERS_DIR = Path(__file__).resolve().parents[3] / "utility" / "checkout"
 if str(HELPERS_DIR) not in sys.path:
     sys.path.insert(0, str(HELPERS_DIR))
