+++ src/api/utility/cart/cart_helpers.py
@@ -9,7 +9,7 @@ import random
 
 import requests
 
-from src.utility.api_test.http_client import request as http_request
+from src.api.utility.http_client import request as http_request
 
 
 # ---------------------------------------------------------------------------
